refactor: raceGame 네이밍 변경
